+++ src/cliargs/constants.lua
@@ -1,0 +1,6 @@
+return {
+  TYPE_COMMAND  = 'command',
+  TYPE_ARGUMENT = 'argument',
+  TYPE_SPLAT    = 'splat',
+  TYPE_OPTION   = 'option',
+}